MEM-01 | up to cmake 3.14 DEMO build switcher. Split project on DEMO and library
